requirements.txt
discord~=2.3.2
twitchAPI~=4.2.0
python-dotenv~=1.0.1